+++ src/jspecview/appletjs/JSVApplet.java
@@ -639,7 +639,7 @@ public class JSVApplet implements JSVAppletInterface,
 	}
 
 	@Override
-	public void processTwoPointGesture(float[][][] touches) {
+	public void processTwoPointGesture(double[][][] touches) {
 		app.vwr.processTwoPointGesture(touches);
 	}
