+++ src/org/jmol/awt/Mouse.java
@@ -402,7 +402,7 @@ class Mouse implements MouseWheelListener, MouseListener,
   private int xWhenPressed, yWhenPressed, modifiersWhenPressed10;
 
   @Override
-  public void processTwoPointGesture(float[][][] touches) {
+  public void processTwoPointGesture(double[][][] touches) {
     // n/a
     
   }
